bot/requirements.txt
discord
boto3
nltk
requests
Pillow